deconstructed FYFD menu value from layout
