+++ src/features/lockYFD/menu/Fund.tsx
@@ -38,7 +38,7 @@ export default function MenuFyfdBalanceFund() {
             </span>
           </fieldset>
         </Box>
-        <Box pb={'1.5em'}>
+        <Box pb={'0.5em'}>
           <Button
             as="button"
             className={styles.buttonSimpleWide}
