+++ src/features/lockYFD/menu/Lock.tsx
@@ -12,7 +12,7 @@ import styles from '@scss/app.module.scss';
 import { selectMyYFD, selectMyFYFD } from '@recoil/connected/balance/selectors';
 import NoticeLoading from '@components/NoticeLoading';
 import { Icons } from '@utilities/variables/icons';
-import LockYfdForm from '../Form';
+import LockYfdForm from './LockForm';
 import { useState } from 'react';
 
 export default function MenuFyfdBalance() {
@@ -48,7 +48,7 @@ export default function MenuFyfdBalance() {
             </span>
           </fieldset>
         </Box>
-        <Box pb={'1.5em'}>
+        <Box pb={'0.5em'}>
           <Button
             as="button"
             variant="outline"
